# openapi/auth.yaml
openapi: 3.0.0
info:
  title: Authentication API
  version: '1.0'
  contact:
    name: support@xata.io
paths:
  /organizations:
    summary: Organizations Management
    description: 'This endpoint enables creating and listing organizations within Xata. Organizations are the top-level entity that contains projects, branches, and team members.'
    get:
      operationId: getOrganizationsList
      summary: Get list of organizations
      description: 'Retrieve the list of all organizations the authenticated user belongs to, including their IDs and names. This endpoint allows users to view all organizations they have access to.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    description: List of organizations the user has access to
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          description: Unique identifier for the organization
                          $ref: "./common/schemas.yaml#/OrganizationID"
                        name:
                          description: Human-readable name of the organization
                          type: string
                        status:
                          description: Current status of the organization
                          $ref: '#/components/schemas/OrganizationStatus'
                      required:
                        - id
                        - name
                        - status
                required:
                  - organizations
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'org:read'
      tags:
        - Organizations
    post:
      operationId: createOrganization
      summary: Create a new organization
      description: Creates a new organization with the authenticated user as its single owner.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'org:write'
      tags:
        - Organizations
  '/organizations/{organizationID}':
    summary: Organization Operations
    description: 'This endpoint provides operations for a specific organization, allowing retrieval, updating, and deletion of an organization by its unique identifier.'
    parameters:
      - $ref: '#/components/parameters/OrganizationIDParam'
    get:
      operationId: getOrganization
      summary: Get organization details
      description: Retrieve detailed information about a specific organization by its ID.
      responses:
        '200':
          description: Organization details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'org:read'
      tags:
        - Organizations
    put:
      operationId: updateOrganization
      summary: Update organization details
      description: 'Update information for an existing organization, such as its name.'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                id:
                  description: Unique identifier for the organization to update
                  $ref: "./common/schemas.yaml#/OrganizationID"
                name:
                  description: New name for the organization
                  type: string
      responses:
        '200':
          description: Organization successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'org:write'
      tags:
        - Organizations
    delete:
      operationId: deleteOrganization
      summary: Delete an organization
      description: '[Not implemented yet] Permanently delete an organization. This action cannot be undone.'
      responses:
        '204':
          description: Organization successfully deleted
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'org:write'
      tags:
        - Organizations
  '/organizations/{organizationID}/api-keys':
    get:
      operationId: listOrganizationAPIKeys
      summary: List API Keys for an organization
      description: Retrieve a list of API keys for a specific organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      responses:
        '200':
          description: List of API keys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    description: Array of API keys for the organization
                    type: array
                    items:
                      $ref: '#/components/schemas/APIKeyPreview'
                required:
                  - keys
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'keys:read'
      tags:
        - API Keys
    post:
      operationId: createOrganizationAPIKey
      summary: Create an Organization API Key
      description: Create a new API key for an organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: API Key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    $ref: '#/components/schemas/FullAPIKey'
                required:
                  - key
        '400':
          description: API key limit reached or invalid request
      security:
        - xata:
            - 'keys:write'
      tags:
        - API Keys
    delete:
      operationId: deleteOrganizationAPIKeys
      summary: Bulk delete API Keys for an organization
      description: Delete multiple API keys in a single request.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  description: Array of API key IDs to delete
                  type: array
                  items:
                    type: string
                  maxItems: 50
              required:
                - ids
      responses:
        '204':
          description: API Keys deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
      security:
        - xata:
            - 'keys:write'
      tags:
        - API Keys
  '/organizations/{organizationID}/members':
    get:
      operationId: listOrganizationMembers
      summary: List members of an organization
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWithID'
                required:
                  - members
      security:
        - xata:
            - 'org:read'
      tags:
        - Organizations
  '/organizations/{organizationID}/members/{userID}':
    delete:
      operationId: removeOrganizationMember
      summary: Remove a member from an organization
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
        - $ref: '#/components/parameters/UserIDParam'
      responses:
        '204':
          description: Member removed
      security:
        - xata:
            - 'org:write'
      tags:
        - Organizations
  '/organizations/{organizationID}/invitations':
    get:
      operationId: listOrganizationInvitations
      summary: List invitations for an organization
      description: 'Retrieve all invitations for an organization with optional filtering by status, email, and name.'
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
        - name: status
          in: query
          description: Filter invitations by status
          schema:
            type: string
            enum:
              - pending
              - expired
        - name: email
          in: query
          description: Filter invitations by email address
          schema:
            type: string
        - name: first_name
          in: query
          description: Filter invitations by first name
          schema:
            type: string
        - name: last_name
          in: query
          description: Filter invitations by last name
          schema:
            type: string
        - name: search
          in: query
          description: Search invitations by email or name
          schema:
            type: string
        - name: first
          in: query
          description: Index of the first result to return (0-based offset for pagination)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: max
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 100
            maximum: 100
            minimum: 1
      responses:
        '200':
          description: List of invitations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrganizationInvitation'
                required:
                  - invitations
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'invite:read'
      tags:
        - Organizations
    post:
      operationId: createOrganizationInvitation
      summary: Send an invitation to join an organization
      description: Send an invitation email to a user to join an organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationInvitationRequest'
      responses:
        '201':
          description: Invitation sent successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        '409':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'invite:write'
      tags:
        - Organizations
  '/organizations/{organizationID}/invitations/{invitationID}':
    get:
      operationId: getOrganizationInvitation
      summary: Get details of a specific invitation
      description: Retrieve detailed information about a specific invitation by its ID.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
        - $ref: '#/components/parameters/InvitationIDParam'
      responses:
        '200':
          description: Invitation details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationInvitation'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'invite:read'
      tags:
        - Organizations
    delete:
      operationId: deleteOrganizationInvitation
      summary: Delete an invitation
      description: Permanently delete an invitation record.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
        - $ref: '#/components/parameters/InvitationIDParam'
      responses:
        '204':
          description: Invitation deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'invite:write'
      tags:
        - Organizations
  '/organizations/{organizationID}/invitations/{invitationID}/resend':
    post:
      operationId: resendOrganizationInvitation
      summary: Resend an invitation
      description: Resend a pending invitation with a fresh expiration time.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
        - $ref: '#/components/parameters/InvitationIDParam'
      responses:
        '204':
          description: Invitation resent successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - 'invite:write'
      tags:
        - Organizations
  /marketplace/register:
    post:
      operationId: registerMarketplace
      summary: Register with a cloud marketplace
      description: Links the authenticated user to a cloud marketplace subscription. Only one marketplace registration is allowed per user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketplaceRegisterRequest'
      responses:
        '200':
          description: Marketplace registration successful
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '409':
          $ref: '#/components/responses/SimpleError'
        '502':
          description: Marketplace registration provider failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
      security:
        - xata:
            - 'marketplace:write'
      tags:
        - Marketplace
      x-excluded: true
...
components:
  parameters:
...
    OrganizationIDParam:
      name: organizationID
      in: path
      required: true
      schema:
        $ref: "./common/schemas.yaml#/OrganizationID"
      description: Unique identifier for a specific organization
...
  schemas:
...
    UserWithID:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          description: Extended user object that includes the unique user identifier
          properties:
            id:
              $ref: '#/components/schemas/UserID'
          required:
            - id
    CreateOrganizationRequest:
      description: Request payload for creating a new organization
      type: object
      properties:
        name:
          description: Name for the new organization
          type: string
      required:
        - name
    Organization:
      description: Organization details including ID and name
      type: object
      properties:
        id:
          description: Unique identifier for the organization
          $ref: "./common/schemas.yaml#/OrganizationID"
        name:
          description: Human-readable name of the organization
          type: string
        status:
          description: Current status of the organization
          $ref: '#/components/schemas/OrganizationStatus'
        marketplace:
          description: 'Marketplace provider for this organization (e.g. "aws"), if billed through a marketplace'
          type: string
          nullable: true
      required:
        - id
        - name
        - status
    OrganizationStatus:
      type: object
      properties:
        status:
          description: 'Indicates whether the organization is active, it''s computed as `!disabled_by_admin AND billing_status == ''ok''`'
          type: string
          enum:
            - enabled
            - disabled
        disabled_by_admin:
          description: Indicates if the organization has been disabled by an admin
          type: boolean
        admin_reason:
          description: Reason for the current admin status
          type: string
        billing_status:
          description: Indicates the status of the organization from a billing perspective
          type: string
          enum:
            - ok
            - no_payment_method
            - invoice_overdue
            - unknown
        billing_reason:
          description: Reason for the current billing status
          type: string
        usage_tier:
          description: 'Usage tier of the organization. t1 is the default for new organizations, t2 is assigned when a valid payment method is on file.'
          type: string
          enum:
            - t1
            - t2
        last_updated:
          description: Timestamp of the last update to the organization's status
          type: string
          format: date-time
        created_at:
          description: Timestamp when the organization was created
          type: string
          format: date-time
      required:
        - status
        - disabled_by_admin
        - billing_status
        - usage_tier
        - last_updated
...
  responses:
    BadRequestError:
      description: Error returned when the request is malformed or contains invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                description: Error identifier for tracking and debugging
                type: string
              message:
                description: Human-readable error message explaining the issue
                type: string
            required:
              - message
    AuthError:
      description: Error returned when authentication or authorization fails
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                description: Error identifier for tracking and debugging
                type: string
              message:
                description: Human-readable error message explaining the issue
                type: string
            example:
              message: invalid API key
            required:
              - message
    SimpleError:
      description: Generic error response
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                description: Error identifier for tracking and debugging
                type: string
              message:
                description: Human-readable error message explaining the issue
                type: string
            required:
              - message
...